@@ -10,25 +10,25 @@
     "fix": "biome lint --write . && biome format --write . && biome check . --write"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.9.0",
+    "@biomejs/biome": "1.9.4",
     "@bitmachina/highlighter": "1.0.0-alpha.6",
-    "@fontsource/fira-mono": "^5.0.14",
+    "@fontsource/fira-mono": "^5.1.0",
     "@json-feed-types/1_1": "^1.0.2",
-    "@rollup/pluginutils": "^5.1.0",
-    "@sentry/esbuild-plugin": "^2.22.2",
-    "@sveltejs/adapter-auto": "^3.2.4",
-    "@sveltejs/adapter-node": "^5.2.2",
-    "@sveltejs/kit": "^2.5.24",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@rollup/pluginutils": "^5.1.3",
+    "@sentry/esbuild-plugin": "^2.22.6",
+    "@sveltejs/adapter-auto": "^3.3.1",
+    "@sveltejs/adapter-node": "^5.2.9",
+    "@sveltejs/kit": "^2.8.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.1",
     "@types/fnv-plus": "^1.3.2",
-    "@types/node": "^20.16.1",
+    "@types/node": "^20.17.7",
     "@types/polka": "^0.5.7",
     "@types/sharedworker": "^0.0.115",
     "dotenv": "^16.4.5",
     "esbuild": "^0.23.1",
     "github-slugger": "^2.0.0",
     "glob": "^10.4.5",
-    "hast-util-to-string": "^3.0.0",
+    "hast-util-to-string": "^3.0.1",
     "just-camel-case": "^6.2.0",
     "mdast-util-to-string": "^4.0.0",
     "mdsvex": "^0.11.2",
@@ -43,56 +43,55 @@
     "remark-math": "^3.0.1",
     "remark-reading-time": "^2.0.1",
     "remark-wiki-link": "^0.0.4",
-    "rollup": "^4.21.0",
+    "rollup": "^4.27.4",
     "rollup-plugin-type-as-json-schema": "^0.2.6",
     "rollup-plugin-visualizer": "^5.12.0",
     "schema-dts": "^1.1.2",
     "sharp": "^0.33.5",
-    "svelte": "^4.2.19",
-    "svelte-check": "^3.8.6",
+    "svelte": "^5.2.7",
+    "svelte-check": "^4.1.0",
     "svelte-seo": "^1.6.1",
-    "sveltekit-html-minifier": "^1.0.3",
+    "sveltekit-html-minifier": "^1.0.4",
     "svgo": "^3.3.2",
-    "tm-grammars": "^1.17.4",
-    "tslib": "^2.7.0",
-    "typescript": "^5.5.4",
+    "tm-grammars": "^1.19.5",
+    "tslib": "^2.8.1",
+    "typescript": "^5.7.2",
     "unified": "^11.0.5",
     "unist-util-visit": "^5.0.0",
     "vite": "^5.4.2",
-    "vite-plugin-dynamic-import": "^1.5.0",
     "vite-plugin-image-optimizer": "^1.1.8",
     "vite-plugin-thumbhash": "^0.1.6",
     "vite-plugin-thumbhash-svg": "workspace:^"
   },
   "type": "module",
   "dependencies": {
-    "@babel/preset-typescript": "^7.24.7",
-    "@codemirror/commands": "^6.6.0",
+    "@babel/preset-typescript": "^7.26.0",
+    "@codemirror/commands": "^6.7.1",
     "@codemirror/lang-javascript": "^6.2.2",
-    "@codemirror/language": "^6.10.2",
-    "@codemirror/lint": "^6.8.1",
+    "@codemirror/language": "^6.10.3",
+    "@codemirror/lint": "^6.8.3",
     "@codemirror/state": "^6.4.1",
-    "@codemirror/view": "^6.33.0",
+    "@codemirror/view": "^6.35.0",
     "@isaacs/ttlcache": "^1.4.1",
     "@lezer/highlight": "^1.2.1",
-    "@octokit/types": "^13.5.0",
+    "@octokit/types": "^13.6.1",
     "@resvg/resvg-js": "^2.6.2",
-    "@sentry/sveltekit": "^8.26.0",
-    "@steeze-ui/svelte-icon": "^1.5.0",
-    "@tabler/icons-svelte": "^3.12.0",
+    "@sentry/sveltekit": "^8.40.0",
+    "@steeze-ui/svelte-icon": "^1.6.2",
+    "@tabler/icons-svelte": "^3.22.0",
     "@tusbar/cache-control": "^1.0.2",
-    "@uiw/codemirror-theme-github": "^4.23.0",
+    "@uiw/codemirror-theme-github": "^4.23.6",
     "Notes": "file:Notes-1.0.0.tgz",
-    "acorn": "^8.12.1",
+    "acorn": "^8.14.0",
     "codemirror": "^6.0.1",
     "fnv-plus": "^1.3.1",
-    "magic-string": "^0.30.11",
+    "magic-string": "^0.30.13",
     "octokit": "^3.2.1",
     "satori": "^0.10.14",
     "slugify": "^1.6.6",
     "super-sitemap": "^0.14.20",
-    "svelte-codemirror-editor": "^1.4.0",
-    "terser": "^5.31.6",
+    "svelte-codemirror-editor": "^1.4.1",
+    "terser": "^5.36.0",
     "xmlbuilder2": "^3.1.1"
   },
   "targets": {
