@@ -5,28 +5,28 @@ import { SENTRY_REPORT_URL } from './src/lib/config.js';
  * @type {import("@sveltejs/kit").CspDirectives}
  */
 const cspDirectives = {
-    'base-uri': ["'self'"],
-    'child-src': ["'self'", "blob:"],
-    'connect-src': ["'self'", "https://*.google-analytics.com", "https://" + SENTRY_HOST],
-    // 'connect-src': ["'self'", 'ws://localhost:*', 'https://hcaptcha.com', 'https://*.hcaptcha.com'],
-    'img-src': ["'self'", 'data:',
+    'base-uri': ["self"],
+    'child-src': ["self", "blob:"],
+    'connect-src': ["self", "https://*.google-analytics.com", "https://" + SENTRY_HOST],
+    // 'connect-src': ["self", 'ws://localhost:*', 'https://hcaptcha.com', 'https://*.hcaptcha.com'],
+    'img-src': ["self", 'data:',
         'https://*.googletagmanager.com'],
-    'font-src': ["'self'", 'data:'],
-    'form-action': ["'self'"],
-    'frame-ancestors': ["'self'"],
+    'font-src': ["self", 'data:'],
+    'form-action': ["self"],
+    'frame-ancestors': ["self"],
     'frame-src': [
-        "'self'",
+        "self",
         // "https://*.stripe.com",
         // "https://*.facebook.com",
         // "https://*.facebook.net",
         // 'https://hcaptcha.com',
         // 'https://*.hcaptcha.com',
     ],
-    'manifest-src': ["'self'"],
-    'media-src': ["'self'", 'data:'],
-    'object-src': ["'none'"],
-    'style-src': ["'self'", "'unsafe-inline'"],
-    // 'style-src': ["'self'", "'unsafe-inline'", 'https://hcaptcha.com', 'https://*.hcaptcha.com'],
+    'manifest-src': ["self"],
+    'media-src': ["self", 'data:'],
+    'object-src': ["none"],
+    'style-src': ["self", "unsafe-inline"],
+    // 'style-src': ["self", "'unsafe-inline'", 'https://hcaptcha.com', 'https://*.hcaptcha.com'],
     'default-src': [
         'self',
         ...(rootDomain ? [rootDomain, `ws://${rootDomain}`] : []),
@@ -42,8 +42,8 @@ const cspDirectives = {
         // 'https://*.sentry.io',
     ],
     'script-src': [
-        'self',
-        'unsafe-inline', // chrome suggestion
+        "self",
+        "unsafe-inline", // chrome suggestion
         'https://*.googletagmanager.com'
         // 'https://*.stripe.com',
         // 'https://*.facebook.com',
@@ -53,9 +53,9 @@ const cspDirectives = {
         // 'https://*.sentry.io',
         // 'https://polyfill.io',
     ],
-    'worker-src': ["'self'", "blob:"],
+    'worker-src': ["self", "blob:"],
     // remove report-to & report-uri if you do not want to use Sentry reporting
-    'report-to': ["'csp-endpoint'"],
+    'report-to': ["csp-endpoint"],
     'report-uri': [
         SENTRY_REPORT_URL,
     ],
